==> Forensics/01_Create_Review_CSV.ipynb <==
# Creation of Reviews CSV File including Bitcoin Conversion
This notebook queries all positive vendor reviews of Berlusconi market from the SQL database and converts the price in euro into Bitcoin in Satoshi. The resulting dataframe is then stored as csv to be easily loadable from other notebooks.

# import needed libraries
import tqdm
import pickle
import pandas as pd
import mysql.connector

# read bitcoin conversion rates file
df_btc = pd.read_csv("BTC_EUR_conversion.csv", delimiter=",")
df_btc.Date = pd.to_datetime(df_btc['Date'])
df_btc.Price = df_btc.Price.str.replace(",", "")
df_btc.Date = df_btc.Date.dt.date
df_btc.head()

# connect to sql database
df_pw = pd.read_csv("pw.csv", delimiter=",")
con = mysql.connector.connect(
    host="localhost",
    user=df_pw["name"][0], 
    passwd=df_pw["password"][0],
    database="DarknetScrape",
)

# get all positive reviews for berlusconi market
cursor = con.cursor()
query = """SELECT price_eur,
`time` ,
customer,
vendor_id,
vendor_review_id

FROM DarknetScrape.vendor_reviews
where market = 'Berlusconi Market'
and positive = 1
;
"""
cursor.execute(query)
results = cursor.fetchall()

df = pd.DataFrame(results, columns=["price_eur","time","customer","vendor_id","vendor_review_id"])
df.head()

# convert the euro price from the reviews into bitcoin in satoshi (value at the time)
df["price_btc_review"] = 0.0
df["price_btc_reviews_satoshi"] = 0.0
for i, row in df.iterrows():
    price_btc_review = float(row.price_eur) / float(df_btc[df_btc["Date"] == row.time.date()]["Price"].iloc[0])
    df.at[i,'price_btc_review'] = price_btc_review
    df.at[i,'price_btc_reviews_satoshi'] = price_btc_review * 100000000
df.head()

# sort by time for more efficient querying in neo4j when looping over reviews
df.time = pd.to_datetime(df.time)
df = df.sort_values(by="time",ascending=True)
df.reset_index(drop=True, inplace=True)

# store results as csv
df.to_csv("reviews_price_btc.csv", index=False)

# check number of reviews
len(df)

display(df[["vendor_review_id","vendor_id","customer","time","price_eur","price_btc_reviews_satoshi"]])
