==> Forensics/03_Payin_and_Escrow_Identification(left_to_right).ipynb <==
# PayIn and Escrow Identification (left to right)
This notebook queries for PayIn transactions matching reviews to vendors that were active on both Berlusconi and Cryptonia Market and have already been identified in previous work. The results were not used during the rest of the project.

import pandas as pd
import mysql.connector
from decimal import Decimal
from neo4j import GraphDatabase
from datetime import timedelta

#Load username and password
df_pw = pd.read_csv("pw.csv", delimiter=",")

#Establish connection to SQL db
df_pw = pd.read_csv("pw.csv", delimiter=",")
con = mysql.connector.connect(
    host="localhost",
    user=df_pw["name"][0], 
    passwd=df_pw["password"][0],
    database="DarknetScrape")
cursor = con.cursor()

#Establish connection to Neo4j
db = GraphDatabase.driver("neo4j://127.0.0.1:7687", auth=(df_pw["name"][0], df_pw["password"][0]))
session = db.session()

#Load known Cryptonia Market vendor addresses from Jochen
j_addresses=pd.read_csv('candidates_final.csv')
j_addresses=j_addresses.drop(columns=['Unnamed: 0'])
j_addresses

#Load all Berlusconi Market reviews
berlusconi_reviews  = pd.read_csv('reviews_price_btc.csv')
berlusconi_reviews.time = pd.to_datetime(berlusconi_reviews.time)

#Query for vendors active on Cryptonia Market from SQL db
query = """select a.vendor_id, b.username from (
SELECT vendor_id, count(*) as count
from vendors
group by vendor_id
) as a
JOIN (SELECT vendor_id, username from vendors where market = 'Cryptonia') b
on a.vendor_id = b.vendor_id
where a.count = 2;
"""
cursor.execute(query)
df = cursor.fetchall()

#Filter for vendors that Jochen already identified
df=df[df['username'].isin(j_addresses['Vendor'])]
#Filter BM reviews for reviews of vendors that Jochen identified
berlusconi_reviews = berlusconi_reviews[berlusconi_reviews["vendor_id"].isin(df["vendor_id"])]
berlusconi_reviews

#Query for payin TAs with change and first level BM escrow addresses & first level mixing TA 
#based on inSum, date and outDegree taken from reviews 
escrow_addresses_v1 = pd.DataFrame(columns=['inDegree', 'outDegree', 'address','vendor'])
for i in berlusconi_reviews.itertuples():
    neo_results = []
    neo_results2 = []
    start = str(i.time.date()- timedelta(days=3))
    end = str(i.time.date())
    minSum = str(round(i.price_btc_reviews_satoshi,0))
    maxSum = str(round(i.price_btc_reviews_satoshi*1.4,0))
    neo_query = session.run("""
    WITH '""" + start + """' as startDate,  '""" + end + """' as endDate, """
    + minSum + """ as minSum, """ + maxSum +""" as maxSum
    MATCH (payin:Transaction) 
        WHERE  t.inDegree = 1 and ( t.outDegree = 2 OR t.outDegree = 1 ) AND date(startDate) <= t.date <= date(endDate) 
        and minSum <= t.inSum<=maxSum
    WITH t, minSum, startDate, endDate
    MATCH (payin)-[r:RECEIVES]->(escrow:Address),
        (payin)-[r1:RECEIVES]->(change:Address)-[s:SENDS]->(payin)
        WHERE minSum  <= r.value <= minSum * 1.05 AND c <> a AND 2000000 < escrow.totalMoneyReceived < 200000000 
        AND escrow.minDate>date('2017-09-08') AND escrow.maxDate<date('2019-07-31')
    WITH a, r, startDate, endDate
    MATCH (a)-[s:SENDS]->(mix:Transaction)
        WHERE date(startDate) <= mix.date <= date(endDate) AND r.value >= s.value 
    RETURN a, r.value
    """)
    for r in neo_query:
        neo_results2.append(r[0])
        neo_results.append(r[1])

    
    df_neo4j2 = pd.DataFrame(neo_results2)
    df_neo4j2["vendor"] = i.vendor_id
    df_neo4j2["review"] = i.vendor_review_id
    df_neo4j2["rvalue"] = pd.Series(neo_results)
    escrow_addresses_v1 = pd.concat([escrow_addresses_v1,df_neo4j2])

escrow_addresses_v1.to_csv("QueryResultsTransaction1EscrowActiveOnBoth0506.csv")    
display(escrow_addresses_v1.head())

escrow_addresses_v1=pd.read_csv("QueryResultsTransaction1EscrowActiveOnBoth0506.csv")

len(escrow_addresses_v1['address'].unique())

#Why inDegree >20 ???
v2=escrow_addresses_v1.groupby(by='address').count().sort_values(by=['inDegree'], ascending=False)
v2=v2[v2['inDegree']>20]
v2=v2.reset_index()
escrow_addresses_v2=v2['address'].tolist()

escrow_addresses_v2_withUsername=escrow_addresses_v1[escrow_addresses_v1['address'].isin(escrow_addresses_v2)]
escrow_addresses_v2_withUsername=escrow_addresses_v2_withUsername[['address', 'vendor']]
escrow_addresses_v2_withUsername=escrow_addresses_v2_withUsername.merge(df, how="inner", left_on='vendor', right_on='vendor_id')
escrow_addresses_v2_withUsername=escrow_addresses_v2_withUsername.drop_duplicates()
escrow_addresses_v2_withUsername

#Query for connection in Neo4j between identified 1st level escrow addresses and Jochen's vendor adresses
#underlying assumption: Only works if the same coin that are spent by the customer are paid out to the vendor
# in other words: If a direct connection between customer and vendor exists in Neo4j
escrow_addresses_v3= pd.DataFrame(columns=['escrow', 'vendor', 'relationships'])
for row in escrow_addresses_v2_withUsername.itertuples():
    vendor=j_addresses[j_addresses['Vendor']==row.username]['Address'].iloc[0]
  
    neo_query = session.run("""
    MATCH p=(escrow:Address {address: '"""+row.address+"""' })-[:RECEIVES*1..5]-(vendor:Address {address: '"""+ vendor+"""'})
    RETURN collect(escrow) as escrow, collect(vendor) as vendor, collect(last(relationships(p))) as relationships""")
    for r in neo_query:
        escrow=r[0]
        vendor=r[1]
        relationships=r[2]
        escrow_addresses_v3.append({'escrow': escrow, 'vendor':vendor, 'relationships':relationships}, ignore_index=True)

display(escrow_addresses_v3.head())
escrow_addresses_v3.to_csv('escrow_addresses_v3.csv')
